Created a single water level DF
Read in the individual water level files and created a pandas-readable dataframe. Unique points can be identified quickly using the "ID" field, or they can be identified using the name field. It's unclear what the "location" field is supposed to indicate (it is always None). This field is kept only because for points outside of Africa it may be important.